autoconf.texi: Don’t say that Darwin is “derived from FreeBSD.”
This is an oversimplification and is probably going to cause
more confusion than it solves.

The commit which added this text was attributed to Sevan Janiyan,
but the text itself was written by me (Zack Weinberg); I revised
the patch but did not remember to also revise the commit message.