tests/local.mk: fix dependencies for generating the wrapper template
Generating the tests wrapper template currently depends on the
dependencies needed for the frozen lib/m4sugar/m4sh.m4f. However,
it also requires that frozen sugar to actually be generated, or the
build will fail if it is not (or at least if it is in the process of
being generated).

Change the dependency to actually depend on the generated file, rather
than on its dependencies.

This kind of failure has been noticed on highly parallel builds on very
loaded machines.

Copyright-paperwork-exempt: yes